Revert "12.05 QuizRepository"
